Seperated configurations into different files
